requirements.txt
fastapi==0.128.0
mistralai==1.10.1
numpy==2.4.1
opencv_python==4.12.0.88
Pillow==12.1.0
pydantic==2.12.5
python-dotenv==1.2.1
timm==1.0.24
torch==2.2.2
torchvision==0.17.2
google-generativeai==0.8.6
